Move legacy Transaction out of core
