+++ client/client-lib/src/mint/mod.rs
@@ -18,7 +18,6 @@ use fedimint_core::modules::mint::{
     BlindNonce, Mint, MintInput, MintOutput, MintOutputBlindSignatures, MintOutputOutcome, Nonce,
     Note,
 };
-use fedimint_core::transaction::legacy::{Input, Output, Transaction};
 use futures::{Future, StreamExt};
 use secp256k1_zkp::{KeyPair, Secp256k1, Signing};
 use serde::{Deserialize, Serialize};
@@ -28,6 +27,7 @@ use tracing::{debug, error, trace, warn};
 
 use crate::api::{GlobalFederationApi, MemberError, OutputOutcomeError};
 use crate::mint::db::{NextECashNoteIndexKey, NotesPerDenominationKey, PendingNotesKey};
+use crate::transaction::legacy::{Input, Output, Transaction};
 use crate::utils::ClientContext;
 use crate::{ChildId, DerivableSecret, FuturesUnordered, MintDecoder};
 
@@ -679,7 +679,6 @@ mod tests {
     use fedimint_core::modules::mint::config::MintClientConfig;
     use fedimint_core::modules::mint::{Mint, MintGen, MintGenParams, MintOutput};
     use fedimint_core::outcome::{SerdeOutputOutcome, TransactionStatus};
-    use fedimint_core::transaction::legacy::Input;
     use fedimint_mint::common::MintDecoder;
     use fedimint_testing::FakeFed;
     use futures::executor::block_on;
@@ -689,6 +688,7 @@ mod tests {
     use crate::api::WsFederationApi;
     use crate::mint::db::NextECashNoteIndexKey;
     use crate::mint::MintClient;
+    use crate::transaction::legacy::Input;
     use crate::{
         module_decode_stubs, BlindNonce, ClientContext, DerivableSecret, TransactionBuilder,
         MINT_SECRET_CHILD_ID,
