+++ client/client-lib/src/transaction.rs
@@ -8,11 +8,176 @@ use fedimint_core::modules::ln::contracts::ContractOutcome;
 use fedimint_core::modules::ln::LightningOutputOutcome;
 use fedimint_core::outcome::legacy::OutputOutcome;
 use fedimint_core::outcome::TransactionStatus;
-use fedimint_core::transaction::legacy::{Input, Output, Transaction};
 use rand::{CryptoRng, RngCore};
 use secp256k1::Secp256k1;
 
+use crate::transaction::legacy::{Input, Output, Transaction};
 use crate::{module_decode_stubs, Client, DecryptedPreimage, MintClient, MintOutputOutcome};
+
+/// Old transaction definition used by old client.
+pub mod legacy {
+    use bitcoin_hashes::Hash;
+    use fedimint_api::core::DynInput;
+    use fedimint_api::core::DynOutput;
+    use fedimint_api::core::{
+        Decoder, LEGACY_HARDCODED_INSTANCE_ID_LN, LEGACY_HARDCODED_INSTANCE_ID_MINT,
+        LEGACY_HARDCODED_INSTANCE_ID_WALLET,
+    };
+    use fedimint_api::encoding::{Decodable, Encodable};
+    use fedimint_api::{ServerModule, TransactionId};
+    use fedimint_core::transaction::{agg_keys, TransactionError};
+    use secp256k1_zkp::{schnorr, XOnlyPublicKey};
+    use serde::{Deserialize, Serialize};
+
+    /// An atomic value transfer operation within the Fedimint system and consensus
+    ///
+    /// The mint enforces that the total value of the outputs equals the total value of the inputs, to prevent creating funds out of thin air. In some cases, the value of the inputs and outputs can both be 0 e.g. when creating an offer to a Lightning Gateway.
+    #[derive(Debug, Clone, Eq, PartialEq, Hash, Encodable, Decodable)]
+    pub struct Transaction {
+        /// [`Input`]s consumed by the transaction
+        pub inputs: Vec<Input>,
+        /// [`Output`]s created as a result of the transaction
+        pub outputs: Vec<Output>,
+        /// Aggregated MuSig2 signature over all the public keys of the inputs
+        pub signature: Option<schnorr::Signature>,
+    }
+
+    /// An Input consumed by a Transaction is defined within a Fedimint Module.
+    ///
+    /// The user must be able to produce an aggregate Schnorr signature for the transaction over all the inputs.
+    ///
+    /// Each input has an associated secret/public key pair.
+    /// Inputs can not have keys if the transaction value is 0. This is useful for non-monetary transactions to announce information to the mint like incoming LN contract offers.
+    #[derive(Debug, Clone, Eq, PartialEq, Hash, Deserialize, Serialize, Encodable, Decodable)]
+    pub enum Input {
+        // TODO: maybe treat every note as a seperate input?
+        Mint(<<fedimint_mint::Mint as ServerModule>::Decoder as Decoder>::Input),
+        Wallet(<<fedimint_wallet::Wallet as ServerModule>::Decoder as Decoder>::Input),
+        LN(<<fedimint_ln::Lightning as ServerModule>::Decoder as Decoder>::Input),
+    }
+
+    // TODO: check if clippy is right
+    #[allow(clippy::large_enum_variant)]
+    #[derive(Debug, Clone, Eq, PartialEq, Hash, Deserialize, Serialize, Encodable, Decodable)]
+    pub enum Output {
+        Mint(<<fedimint_mint::Mint as ServerModule>::Decoder as Decoder>::Output),
+        Wallet(<<fedimint_wallet::Wallet as ServerModule>::Decoder as Decoder>::Output),
+        LN(<<fedimint_ln::Lightning as ServerModule>::Decoder as Decoder>::Output),
+    }
+
+    impl Transaction {
+        /// Hash of the transaction (excluding the signature).
+        ///
+        /// Transaction signature commits to this hash.
+        /// To generate it without already having a signature use [`Self::tx_hash_from_parts`].
+        pub fn tx_hash(&self) -> TransactionId {
+            Self::tx_hash_from_parts(&self.inputs, &self.outputs)
+        }
+
+        /// Generate the transaction hash.
+        pub fn tx_hash_from_parts(inputs: &[Input], outputs: &[Output]) -> TransactionId {
+            let erased_inputs = inputs
+                .iter()
+                .map(|input| match input.clone() {
+                    Input::Mint(i) => DynInput::from_typed(LEGACY_HARDCODED_INSTANCE_ID_MINT, i),
+                    Input::Wallet(i) => {
+                        DynInput::from_typed(LEGACY_HARDCODED_INSTANCE_ID_WALLET, i)
+                    }
+                    Input::LN(i) => DynInput::from_typed(LEGACY_HARDCODED_INSTANCE_ID_LN, i),
+                })
+                .collect::<Vec<DynInput>>();
+            let erased_outputs = outputs
+                .iter()
+                .map(|output| match output.clone() {
+                    Output::Mint(o) => DynOutput::from_typed(LEGACY_HARDCODED_INSTANCE_ID_MINT, o),
+                    Output::Wallet(o) => {
+                        DynOutput::from_typed(LEGACY_HARDCODED_INSTANCE_ID_WALLET, o)
+                    }
+                    Output::LN(o) => DynOutput::from_typed(LEGACY_HARDCODED_INSTANCE_ID_LN, o),
+                })
+                .collect::<Vec<DynOutput>>();
+
+            let mut engine = TransactionId::engine();
+            erased_inputs
+                .consensus_encode(&mut engine)
+                .expect("write to hash engine can't fail");
+            erased_outputs
+                .consensus_encode(&mut engine)
+                .expect("write to hash engine can't fail");
+            TransactionId::from_engine(engine)
+        }
+
+        /// Validate the aggregated Schnorr Signature signed over the tx_hash
+        pub fn validate_signature(
+            &self,
+            keys: impl Iterator<Item = XOnlyPublicKey>,
+        ) -> Result<(), TransactionError> {
+            let keys = keys.collect::<Vec<_>>();
+
+            // If there are no keys from inputs there are no inputs to protect from re-binding. This
+            // behavior is useful for non-monetary transactions that just announce something, like LN
+            // incoming contract offers.
+            if keys.is_empty() {
+                return Ok(());
+            }
+
+            // Unless keys were empty we require a signature
+            let signature = self
+                .signature
+                .as_ref()
+                .ok_or(TransactionError::MissingSignature)?;
+
+            let agg_pub_key = agg_keys(&keys);
+            let msg = secp256k1_zkp::Message::from_slice(&self.tx_hash()[..])
+                .expect("hash has right length");
+
+            if secp256k1_zkp::global::SECP256K1
+                .verify_schnorr(signature, &msg, &agg_pub_key)
+                .is_ok()
+            {
+                Ok(())
+            } else {
+                Err(TransactionError::InvalidSignature)
+            }
+        }
+
+        pub fn into_type_erased(self) -> fedimint_core::transaction::Transaction {
+            fedimint_core::transaction::Transaction {
+                inputs: self
+                    .inputs
+                    .into_iter()
+                    .map(|input| match input {
+                        Input::Mint(input) => {
+                            DynInput::from_typed(LEGACY_HARDCODED_INSTANCE_ID_MINT, input)
+                        }
+                        Input::Wallet(input) => {
+                            DynInput::from_typed(LEGACY_HARDCODED_INSTANCE_ID_WALLET, input)
+                        }
+                        Input::LN(input) => {
+                            DynInput::from_typed(LEGACY_HARDCODED_INSTANCE_ID_LN, input)
+                        }
+                    })
+                    .collect(),
+                outputs: self
+                    .outputs
+                    .into_iter()
+                    .map(|output| match output {
+                        Output::Mint(output) => {
+                            DynOutput::from_typed(LEGACY_HARDCODED_INSTANCE_ID_MINT, output)
+                        }
+                        Output::Wallet(output) => {
+                            DynOutput::from_typed(LEGACY_HARDCODED_INSTANCE_ID_WALLET, output)
+                        }
+                        Output::LN(output) => {
+                            DynOutput::from_typed(LEGACY_HARDCODED_INSTANCE_ID_LN, output)
+                        }
+                    })
+                    .collect(),
+                signature: self.signature,
+            }
+        }
+    }
+}
 
 pub trait Final {
     fn is_final(&self) -> bool;
