+++ integrationtests/tests/tests.rs
@@ -13,7 +13,6 @@ use fedimint_mint::{MintConsensusItem, MintOutputSignatureShare};
 use fedimint_server::consensus::TransactionSubmissionError::TransactionError;
 use fedimint_server::consensus::TransactionSubmissionError::TransactionReplayError;
 use fedimint_server::epoch::ConsensusItem;
-use fedimint_server::transaction::legacy::Output;
 use fedimint_server::transaction::TransactionError::UnbalancedTransaction;
 use fedimint_wallet::PegOutSignatureItem;
 use fedimint_wallet::WalletConsensusItem::PegOutSignature;
@@ -21,6 +20,7 @@ use fixtures::{rng, secp, sha256};
 use futures::future::{join_all, Either};
 use mint_client::logging::LOG_TEST;
 use mint_client::mint::MintClient;
+use mint_client::transaction::legacy::Output;
 use mint_client::transaction::TransactionBuilder;
 use mint_client::{ClientError, ConfigVerifyError};
 use threshold_crypto::{SecretKey, SecretKeyShare};
